chore: split pages and app routes for nextjs
