+++ packages/parsers/src/index.ts
@@ -3,7 +3,8 @@
  */
 
 export * from "./http-format";
-export * from "./nextjs/nextjs-parser";
+export * from "./next-app/next-app-parser";
+export * from "./next-pages/next-pages-parser";
 export * from "./nestjs/nestjs-parser";
 export * from "./trpc/trpc-parser";
 export * from "./shared/zod-schema-parser";
